pyproject.toml
[tool.poetry]
name = "beehive"
version = "0.1.0"
description = ""
authors = ["Christian <christian@baton.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-exporter-otlp = "^1.22.0"
opentelemetry-instrumentation = "^0.43b0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
